Setup - Complete - Created - Signup form
The setup for the react project is complete. along with all the setup imports of the technologies used - vite setup, appwrite signup, tailwind import, shadcn import.

The file structure of the project is complete and basic components have been created. along with that signup page is created. it took 4+ hours in total. learning and work in progress.